add IDiagnosticProjection inside onData Method
changed also the view linked with it

+++ views/dashboard/DashboardIndex.tsx
@@ -8,34 +8,38 @@ import DashboardViewModel from "../../scripts/DashboardViewModel";
 export default class DashboardIndex extends View<DashboardViewModel> {
 
     render() {
-        let projections = _.map(this.viewModel.model, (value: any, key: string) => {
-            return <ProjectionPanel title={key} projection={value}
-                                    stop={() => this.viewModel.stop(key)}
-                                    pause={() => this.viewModel.pause(key)}
-                                    resume={() => this.viewModel.resume(key)}
-                                    saveSnapshot={() => this.viewModel.saveSnapshot(key)}
-                                    deleteSnapshot={() => this.viewModel.deleteSnapshot(key)}/>
-        });
+        let projections = [];
+        if(this.viewModel.model) ///model is ReadyState
+            projections = _.map(this.viewModel.model.list, (value: any, key: string) => {
+                return <ProjectionPanel title={key} projection={value}
+                                        stop={() => this.viewModel.stop(key)}
+                                        pause={() => this.viewModel.pause(key)}
+                                        resume={() => this.viewModel.resume(key)}
+                                        saveSnapshot={() => this.viewModel.saveSnapshot(key)}
+                                        deleteSnapshot={() => this.viewModel.deleteSnapshot(key)}/>
+            });
 
         return (
             <div>
-                <PageHeader>Projections <small>Overview</small></PageHeader>
-                    <table className="table table-striped">
-                        <thead>
-                        <tr>
-                            <th>Projection Name</th>
-                            <th>Status</th>
-                            <th>Size</th>
-                            <th>Events</th>
-                            <th>ReadModels</th>
-                            <th>Actions</th>
-                            <th>Snapshot</th>
-                        </tr>
-                        </thead>
-                        <tbody>
-                            {projections}
-                        </tbody>
-                    </table>
+                <PageHeader>Projections
+                    <small>Overview</small>
+                </PageHeader>
+                <table className="table table-striped">
+                    <thead>
+                    <tr>
+                        <th>Projection Name</th>
+                        <th>Status</th>
+                        <th>Size</th>
+                        <th>Events</th>
+                        <th>ReadModels</th>
+                        <th>Actions</th>
+                        <th>Snapshot</th>
+                    </tr>
+                    </thead>
+                    <tbody>
+                    {projections}
+                    </tbody>
+                </table>
             </div>
         );
     }
